test(analysis): inline detector and common_state unit tests
Move tests from tests/test_detector.rs and tests/test_common_state.rs into
#[cfg(test)] mod tests blocks of src/analysis/detector.rs and
src/analysis/common_state.rs. Suite count holds at 381 tests.
